improvement: split section
