+++ apps/kg/modules/leaderboard/module.tsx
@@ -1,5 +1,5 @@
 import React, { ReactElement } from 'react';
-import { ContentSection } from './section/content-section';
+import { ContentSection } from './section/module';
 import { BaseLayout } from '../common';
 import { BreadCrumb } from '@mknows-frontend-services/components/atoms';
 import { leaderBoardBreadCumbs } from '../administration/constant';
