+++ apps/kg/modules/leaderboard/section/ranking-bottom.tsx
